+++ CHANGELOG.md
@@ -6,6 +6,16 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 
+## [0.54.0] - 2023-11-xx
+
+### Added
+
+- Add option to set the display units on a Y-axis to units such as Thousands or
+  Millions via the [`ChartAxis::set_display_unit_type()`] method.
+
+[`ChartAxis::set_display_unit_type()`]: https://docs.rs/rust_xlsxwriter/latest/rust_xlsxwriter/struct.ChartAxis.html#method.set_display_unit_type
+
+
 ## [0.53.0] - 2023-10-30
 
 ### Added